apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calcio-platform-api"
version = "0.1.0"
description = "Sprint 0 backend foundation for Calcio Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.16,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.2,<3.0",
  "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.7,<1.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
